Upload the first 3 files of this project
They are now 3 files on this project. conf.php contains environment variables that are needed in others pages of the Dashboard.
add.php is the first page, that allow you to add a shared folder to your Samba server. The interface is simply, but it's a first version.
edit.php is not finished, i'm working on it.